+++ integration_tests/juniper_tests/src/lib.rs
@@ -7,6 +7,8 @@ mod custom_scalar;
 #[cfg(test)]
 mod explicit_null;
 #[cfg(test)]
+mod infallible_as_field_error;
+#[cfg(test)]
 mod issue_371;
 #[cfg(test)]
 mod issue_398;